requirements.txt
obs-websocket-py
Flask==1.0.2
requests==2.18.4
dialogflow==0.4.0
python-dotenv==0.8.2
pusher
multiprocess
speechrecognition
pyttsx3
pyaudio
pipwin
